reorganize files by folder
